# wiretest/wiredemo/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "wire demo API",
        "contact": {},
        "version": "1.0.0"
    },
    "paths": {
        "/hzw": {
            "get": {
                "description": "根据ID查询Hzw对象",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hzw"
                ],
                "summary": "根据ID查询Hzw",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Hzw ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Hzw"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "创建新的Hzw对象",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hzw"
                ],
                "summary": "创建Hzw",
                "parameters": [
                    {
                        "description": "Hzw信息",
                        "name": "hzw",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Hzw"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Hzw"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/hzw2": {
            "get": {
                "description": "根据ID查询Hzw2对象",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hzw"
                ],
                "summary": "根据ID查询Hzw2",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Hzw2 ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wiredemo_internal_repository_model.Hzw2"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "创建新的Hzw2对象",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hzw"
                ],
                "summary": "创建Hzw2",
                "parameters": [
                    {
                        "description": "Hzw2信息",
                        "name": "hzw",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wiredemo_internal_repository_model.Hzw2"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wiredemo_internal_repository_model.Hzw2"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/hzwtxtest": {
            "put": {
                "description": "创建新的Hzw对象",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hzw"
                ],
                "summary": "创建Hzw",
                "parameters": [
                    {
                        "description": "Hzw信息",
                        "name": "hzw",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Hzw"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Hzw"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Hzw": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "createdAt": {
                    "description": "CreatedAt is a field that contains create time",
                    "type": "string"
                },
                "decimal1": {
                    "type": "number"
                },
                "id": {
                    "description": "gorm.Model      // this is a struct that contains Id, CreatedAt, UpdatedAt, DeletedAt",
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "time1": {
                    "description": "AUTOUPDATETIME means update time when update, like UpdatedAt",
                    "type": "string"
                },
                "time2": {
                    "description": "AUTOCREATETIME means create time when insert, like CreatedAt",
                    "type": "string"
                },
                "time3": {
                    "type": "string"
                },
                "updatedAt": {
                    "description": "UpdatedAt is a field that contains update time",
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "wiredemo_internal_repository_model.Hzw2": {
            "type": "object",
            "properties": {
                "Age": {
                    "type": "integer"
                },
                "CreatedAt": {
                    "type": "string"
                },
                "Decimal1": {
                    "type": "number"
                },
                "Id": {
                    "type": "integer"
                },
                "Name": {
                    "type": "string"
                },
                "Time1": {
                    "type": "string"
                },
                "Time2": {
                    "type": "string"
                },
                "Time3": {
                    "type": "string"
                },
                "UpdatedAt": {
                    "type": "string"
                },
                "Version": {
                    "type": "integer"
                }
            }
        }
    }
}