# wiretest/wiredemo/internal/repository/model/Hzw2.go
// Code generated by aic-ent, DO NOT EDIT
// Code generated by aic-ent, DO NOT EDIT
// Code generated by aic-ent, DO NOT EDIT

package model
import (
	"time"
)
 

// Hzw2 mapped from table HZW2
type Hzw2 struct {
   Id  uint `gorm:"column:ID;primaryKey;not null" json:"Id"`
   Name  string `gorm:"column:NAME;not null;index:HZW2_NAME,priority:1" json:"Name"`
   Age  int `gorm:"column:AGE" json:"Age"`
   Version  int32 `gorm:"column:VERSION" json:"Version"`
   CreatedAt  time.Time `gorm:"column:CREATED_AT" json:"CreatedAt"`
   UpdatedAt  time.Time `gorm:"column:UPDATED_AT" json:"UpdatedAt"`
   Time1  time.Time `gorm:"AUTOUPDATETIME;column:TIME1" json:"Time1"`
   Time2  time.Time `gorm:"AUTOCREATETIME;column:TIME2" json:"Time2"`
   Time3  time.Time `gorm:"column:TIME3" json:"Time3"`
   Decimal1  float64 `gorm:"column:DECIMAL1" json:"Decimal1"`
}

// TableName HZW2 table name
func (*Hzw2) TableName() string {
   return "HZW2"
}